# [wip] kvserver: add avgProposalToLocalApplicationLatency