--- a/pkg/kv/kvserver/closedts/multiregion/BUILD.bazel
+++ b/pkg/kv/kvserver/closedts/multiregion/BUILD.bazel
@@ -1,0 +1,33 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
+
+go_library(
+    name = "multiregion",
+    srcs = ["policy_refresher.go"],
+    importpath = "github.com/cockroachdb/cockroach/pkg/kv/kvserver/closedts/multiregion",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//pkg/clusterversion",
+        "//pkg/kv/kvserver/closedts",
+        "//pkg/roachpb",
+        "//pkg/settings/cluster",
+        "//pkg/util/stop",
+        "//pkg/util/timeutil",
+    ],
+)
+
+go_test(
+    name = "multiregion_test",
+    srcs = ["policy_refresher_test.go"],
+    embed = [":multiregion"],
+    deps = [
+        "//pkg/kv/kvserver/closedts",
+        "//pkg/roachpb",
+        "//pkg/settings/cluster",
+        "//pkg/testutils",
+        "//pkg/util/leaktest",
+        "//pkg/util/log",
+        "//pkg/util/stop",
+        "@com_github_cockroachdb_errors//:errors",
+        "@com_github_stretchr_testify//require",
+    ],
+)