--- a/pkg/kv/kvserver/closedts/policy.go
+++ b/pkg/kv/kvserver/closedts/policy.go
@@ -8,7 +8,7 @@ package closedts
 import (
 	"time"
 
-	"github.com/cockroachdb/cockroach/pkg/roachpb"
+	"github.com/cockroachdb/cockroach/pkg/kv/kvserver/closedts/ctpb"
 	"github.com/cockroachdb/cockroach/pkg/util/hlc"
 )
 
@@ -104,14 +104,14 @@ func TargetForPolicy(
 	lagTargetDuration time.Duration,
 	leadTargetOverride time.Duration,
 	sideTransportCloseInterval time.Duration,
-	policy roachpb.RangeClosedTimestampPolicy,
+	policy ctpb.LatencyBasedRangeClosedTimestampPolicy,
 ) hlc.Timestamp {
 	var leadTimeAtSender time.Duration
 	switch policy {
-	case roachpb.LAG_BY_CLUSTER_SETTING:
+	case ctpb.LAG_BY_CLUSTER_SETTING:
 		// Simple calculation: lag now by desired duration.
 		leadTimeAtSender = -lagTargetDuration
-	case roachpb.LEAD_FOR_GLOBAL_READS:
+	case ctpb.LEAD_FOR_GLOBAL_READS_WITH_NO_LATENCY_INFO:
 		// Override entirely with cluster setting, if necessary.
 		if leadTargetOverride != 0 {
 			leadTimeAtSender = leadTargetOverride