--- a/pkg/kv/kvserver/closedts/policy_calculation.go
+++ b/pkg/kv/kvserver/closedts/policy_calculation.go
@@ -7,10 +7,26 @@ package closedts
 
 import (
 	"fmt"
+	"math"
 	"time"
 
 	"github.com/cockroachdb/cockroach/pkg/kv/kvserver/closedts/ctpb"
 )
+
+// FindBucketBasedOnNetworkRTT maps a network RTT to a closed timestamp policy bucket.
+// It divides RTT by policy interval, adds 1 for zero-based indexing, and offsets by
+// the base policy enum value.
+func FindBucketBasedOnNetworkRTT(
+	networkRTT time.Duration,
+) ctpb.LatencyBasedRangeClosedTimestampPolicy {
+	if networkRTT < 0 {
+		return ctpb.LEAD_FOR_GLOBAL_READS_WITH_NO_LATENCY_INFO
+	}
+	bucketNum := int32(math.Floor(float64(networkRTT)/float64(closedTimestampPolicyLatencyInterval))) + 1
+	policy := ctpb.LatencyBasedRangeClosedTimestampPolicy(
+		bucketNum + int32(ctpb.LEAD_FOR_GLOBAL_READS_WITH_NO_LATENCY_INFO))
+	return min(ctpb.LEAD_FOR_GLOBAL_READS_LATENCY_EQUAL_OR_GREATER_THAN_300MS, policy)
+}
 
 // computeNetworkRTTBasedOnPolicy converts a closed timestamp policy to an estimated
 // network RTT. Returns default max RTT for no latency info, or midpoint RTT for