--- a/pkg/kv/kvserver/closedts/setting.go
+++ b/pkg/kv/kvserver/closedts/setting.go
@@ -49,3 +49,23 @@ var LeadForGlobalReadsOverride = settings.RegisterDurationSetting(
 	settings.NonNegativeDuration,
 	settings.WithPublic,
 )
+
+// LeadForGlobalReadsAutoTuneInterval controls how frequently the system adjusts closed
+// timestamp policies for ranges based on network latency measurements. When enabled,
+// ranges are assigned policies based on observed latency between their leaseholder and furthest follower.
+// This dynamic adjustment helps optimize closed timestamp lead times. If no latency
+// data is available, falls back to default lead times. Note that the
+// LeadForGlobalReadsOverride setting takes precedence if set.
+var LeadForGlobalReadsAutoTuneInterval = settings.RegisterDurationSetting(
+	settings.SystemVisible,
+	"kv.closed_timestamp.lead_for_global_reads_auto_tune_interval",
+	"interval at which the system updates closed timestamp policies for ranges based on observed network latency. "+
+		"If non-zero, global-read ranges are periodically assigned closed timestamp policies based on their "+
+		"leaseholder-to-furthest-follower latencies, affecting their closed timestamp lead times. If "+
+		"kv.closed_timestamp.lead_for_global_reads_override is set, "+
+		"it takes precedence over the auto-tuned value. Set to 0 to disable auto-tuning and use default lead time for "+
+		"global read ranges without using observed network latency.",
+	5*time.Minute,
+	settings.NonNegativeDuration,
+	settings.WithPublic,
+)