--- a/pkg/kv/kvserver/replica.go
+++ b/pkg/kv/kvserver/replica.go
@@ -7,7 +7,7 @@ package kvserver
 
 import (
 	"context"
-	fmt "fmt"
+	"fmt"
 	"slices"
 	"sync"
 	"sync/atomic"