… docs/generated/settings/settings-for-tenants.txt
… docs/generated/settings/settings.html
… pkg/BUILD.bazel
… pkg/kv/kvserver/closedts/multiregion/policy_refresher_test.go
… pkg/kv/kvserver/closedts/policy_calculation_test.go
… pkg/kv/kvserver/closedts/policy_test.go
… pkg/kv/kvserver/closedts/sidetransport/BUILD.bazel
… pkg/kv/kvserver/closedts/sidetransport/receiver.go
… pkg/kv/kvserver/closedts/sidetransport/receiver_test.go
… pkg/kv/kvserver/closedts/sidetransport/sender.go
… pkg/kv/kvserver/closedts/sidetransport/sender_helper_test.go
… pkg/kv/kvserver/closedts/sidetransport/sender_test.go
… pkg/kv/kvserver/helpers_test.go
… pkg/kv/kvserver/replica_closedts.go
… pkg/kv/kvserver/replica_closedts_test.go
… pkg/kv/kvserver/replica_init.go
… pkg/kv/kvserver/replica_proposal_buf_test.go
… pkg/kv/kvserver/store.go
… pkg/kv/txn.go
… pkg/server/server.go